model/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
transformers==4.44.2
torch>=2.2.0
peft==0.12.0
boto3==1.35.8
python-multipart==0.0.9
orjson==3.10.7
httpx==0.27.2
pytest==8.2.1
tenacity==8.5.0
python-dotenv==1.0.1
prometheus-client==0.20.0
psutil==5.9.8